cricket_vision_prototype/requirements.txt
numpy>=1.26
opencv-python>=4.9
PyYAML>=6.0
scipy>=1.11
